[feature]增加普通/会员列表分页加载，修复今日操作规划列表接口调整引起的bug

+++ app/src/main/java/com/yunkahui/datacubeper/common/bean/DesignSub.java
@@ -5,6 +5,6 @@ public interface DesignSub {
     String getPlan_type();
     String getBankCardName();
     long getDate();
-    double getAmount();
+    int getAmount();
     String getOperation();
 }
